minor opt tab refactor (#2062)
