+++ libs/gi/db-ui/src/hooks/index.ts
@@ -10,4 +10,5 @@ export * from './useEquippedInTeam'
 export * from './useOptConfig'
 export * from './useTeam'
 export * from './useTeamChar'
+export * from './useTeammateArtifactIds'
 export * from './useWeapon'
